rag-backend/requirements.txt
boto3>=1.26.0
aws-lambda-powertools>=2.10.0
python-dotenv>=1.0.0
